[Sanitizers] Avoid strlen in regexec interceptor if REG_STARTEND is used
REG_STARTEND allows the specified string to contain embedded NUL bytes
and for it to not be NUL-terminated. This is an extension that is not
POSIX, but is available on many platforms. If a platform doesn't support
REG_STARTEND but still has SANATIZER_INTERCEPT_REGEX defined, this
should still continue to work as expected.

This commit intentionally does not attempt to implement the range read
validation, as doing so requires a definition of sanitizer_regmatch_t
which is slightly difficult to achieve, so it will be done in a later
commit.